requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-groq>=0.0.2
groq>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.15.1
python-docx>=0.8.11
python-pptx>=0.6.21
numpy>=1.24.0
tiktoken>=0.5.1
pandas>=2.0.0
unstructured>=0.10.0
pdf2image>=1.16.3
flask>=2.3.3
python-dotenv>=1.0.0
pyyaml>=6.0
matplotlib>=3.7.0
flask-session>=0.5.0
PyMuPDF>=1.21.1
Pillow>=9.0.0 